Add save feature and improve button layout
The changes in this commit include adding save functionality within ConnectionView.vue and enhancing the layout of buttons in App.vue.

The save feature was added by defining a new 'emit' function that emits a 'save' event with a train object as input. This event is then fired when the save button is clicked within each train's information panel.

The layout changes in App.vue were made to improve visual clarity. The 'Add new itinerary' and 'Refresh' buttons now have margins, which help in separating these buttons from other elements on the page. This should make the application easier to interact with for users.